requirements.txt
autopep8==2.0.0
azure-core==1.26.0
azure-eventhub==5.10.1
certifi==2022.9.24
charset-normalizer==2.1.1
Faker==15.1.1
idna==3.4
mysql-connector==2.2.9
psycopg2==2.9.5
pycodestyle==2.9.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
uamqp==1.6.3
urllib3==1.26.12
